updated Length column

+++ BarManagerApplication.Backend/src/BackendGestionaleBar.Database/dbo/Tables/Images.sql
@@ -2,7 +2,7 @@
 (
 	[Id] UNIQUEIDENTIFIER NOT NULL DEFAULT newid(),
 	[Path] NVARCHAR(256) NOT NULL,
-	[Length] INTEGER NOT NULL,
+	[Length] BIGINT NOT NULL,
 	[CreatedDate] DATETIME NOT NULL,
 	[LastModifiedDate] DATETIME NULL,
